[330. Patching Array][Accepted]committed by LeetCode Extension

+++ 330-Patching-Array/solution.java
@@ -2,8 +2,8 @@ public class Solution {
     public int minPatches(int[] nums, int n) {
         int res = 0;
         int i = 0;
-        int bound = 1;
-        int count = 0;
+        long bound = 1;
+        long count = 0;
         while (bound <= n) {
             if (i < nums.length && nums[i] <= bound && nums[i] + count >= bound) {
                 count += nums[i++];
